@@ -11,25 +11,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~14.2.2",
-    "@angular/cdk": "^13.3.9",
-    "@angular/common": "~14.2.2",
-    "@angular/compiler": "~14.2.2",
-    "@angular/core": "~13.3.11",
+    "@angular/animations": "^14.2.2",
+    "@angular/cdk": "^14.2.2",
+    "@angular/common": "^14.2.2",
+    "@angular/compiler": "^14.2.2",
+    "@angular/core": "^14.2.2",
     "@angular/flex-layout": "^10.0.0-beta.32",
-    "@angular/forms": "~14.2.2",
-    "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "~14.2.2",
-    "@angular/platform-browser-dynamic": "~14.2.2",
-    "@angular/router": "~14.2.2",
+    "@angular/forms": "^14.2.2",
+    "@angular/material": "^14.2.2",
+    "@angular/platform-browser": "^14.2.2",
+    "@angular/platform-browser-dynamic": "^14.2.2",
+    "@angular/router": "^14.2.2",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.9",
-    "@angular/cli": "~13.3.9",
-    "@angular/compiler-cli": "~14.2.2",
+    "@angular-devkit/build-angular": "^14.2.3",
+    "@angular/cli": "^14.2.3",
+    "@angular/compiler-cli": "^14.2.2",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
